requirements.txt
torch
torchvision
uv
matplotlib
networkx
tqdm
pyyaml
tensorboard
huggingface-hub
